Put generic tree-sitter parsing entrypoing into /scripts
