+++ tests/list_and_option/Makefile
@@ -1,10 +1,10 @@
 #TODO: json_reader.ml should be an autogenerated json_reader.re at some point
-all: grammar.json ex1.json ast.re json_reader.ml
+all: grammar.json ex1.json AST.ml json_reader.ml
 
 TOP=../..
 
 # you need 'make' from $(TOP)
-GEN=$(TOP)/_build/default/bin/main_codegen.exe
+GEN=$(TOP)/bin/ocaml-tree-sitter
 # you need 'npm install tree-sitter-cli' from $(TOP)
 CLI=$(TOP)/node_modules/.bin/tree-sitter
 
@@ -15,21 +15,19 @@ grammar.json: grammar.js
 ex1.json: ex1.txt grammar.json
 	npm init -y
 	npm install
-	./dumper.js ex1.txt > ex1.json
+	cp $(TOP)/scripts/tree-sitter-parser.js .
+	# ydump is the json pretty-printer that comes with atdgen
+	./tree-sitter-parser.js ex1.txt | ydump > ex1.json
 
-ast.re: grammar.json
-	$(GEN) -codegen_types $^ > $@
-
-#TODO intermediate2 error in generated file for now
-#json_reader.re: grammar.json
-#	$(GEN) -codegen_jsonreader $^ > $@
-
+AST.ml: grammar.json
+	$(GEN) $^
 
 clean:
 	rm -f *.gyp index.js
 	rm -rf src/ build/
 	rm -f package-lock.json package.json
 	rm -f .merlin
+	rm -f AST.ml Parse.ml
 
 # take care, because dune runtest from $(TOP) will not work without those files
 distclean:
